Update PG_admin_Covid_19
This SQL project is a Data Analysis exercise which focuses on breaking down the Covid-19 Data coming from Our World In Data. The SQL script focuses on creating a total of 5 tables from the main database which best describe the most important information from the data. This information includes:
1. The total cases and deaths from respective countries and continents
2. The infection, case fatality and death rates
3. The highest recorded ICU admissions
4. The total tests and vaccinations over time for certain interest countries

These tables can be found under the folder 'Summaries' in this repository.